pyproject.toml
[project]
name = "pyapi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "~=3.12"
dependencies = [
    "libsql-client>=0.3.1",
    "python-dotenv>=1.0.1",
    "sanic-ext>=23.12.0",
    "sanic>=24.6.0",
    "aioquic>=1.2.0",
    "pyjwt>=2.10.1",
    "pytimeparse2>=1.7.1",
    "bcrypt>=4.2.1",
]

